Renamed TreeContextMenu to ContextMenu
