+++ src/Field/FieldElements.tsx
@@ -9,7 +9,7 @@ import { useAppSelector, useAppDispatch } from "../Store/hooks";
 import { AppDispatch } from "../Store/store";
 import { selectPathByValidId } from "../Navbar/pathsSlice";
 import { selectHoveredWaypointIds, selectSelectedWaypointIds, itemSelected, ItemType, itemMouseEnter, itemMouseLeave, selectSelectedSplineIds, selectHoveredSplineIds, splineSelected, splineMouseEnter, splineMouseLeave } from "../Tree/tempUiSlice";
-import { MenuLocation, WaypointContextMenu, SplineContextMenu } from "../Tree/TreeContextMenu";
+import { MenuLocation, WaypointContextMenu, SplineContextMenu } from "../Tree/ContextMenu";
 import { selectActiveRoutine, selectHiddenWaypointIds } from "../Tree/uiSlice";
 import { isControlWaypoint, waypointMoved, waypointRobotRotated, MagnitudePosition, waypointMagnitudeMoved, selectWaypointById, ControlWaypoint } from "../Tree/waypointsSlice";
 import { Point, makeCurve, parameterRange, INCH, DEGREE, FEET, flatten, makePointFromKonvaEvent, makePointFromPolar } from "./mathUtils";
